WORKAROUND: If the console output at AppVeyor becomes too long, then an error "Log entry string is too long. A string written to the event log cannot exceed 32766 characters" is generated, e.g.
https://ci.appveyor.com/project/HenrikBengtsson/matrixstats/build/1.0.107

This error would in turn, via 'on_failure:' rule, trigger all log files to
be dumped to the console make the console output ridicolously long and
taking forever for the AppVeyor page to render.

The workaround is to ZIP all *.Rout files; this will free up almost
equally number of console output lines as there are *.Rout files, which
hopefully keeps the total number of character below the above threshhold.